SELinux: Implement newer policydb features (version 30) into Kernel 3.0.8
Android 8.1 uses an advanced SELinux policy strategy, which requires a higher version of policydb (>= 30).
But Kernel 3.0.8 is too old to serve its requirements, and updating kernel to newer version is really a huge work corresponding to total source upgrade.
So the moderate method is to implement those newer features of version 30 here.